Start rewriting this addin
