+++ .vscode/launch.json
@@ -1,0 +1,17 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Python: Attach",
+            "type": "python",
+            "request": "attach",
+            "pathMappings": [  {
+                "localRoot": "${workspaceRoot}",
+                "remoteRoot": "${workspaceRoot}"}],
+            "osx": {"filePath":"${file}"},
+            "windows": {"filePath":"${file}"},
+            "port": 9000,
+            "host": "localhost"
+        }
+    ]
+}
